Use `class Peekable` for parser (#228)
This pull request adds a `class Peekable` which augments an argument
`Iterator` with a `peek` method. The parser now takes as input an
instance of `class Peekable` that produces a series of tokens rather
than relying on a materialised `list` of tokens. For right now,
`tokenize` is altered to wrap the returned `list` of tokens in an
instance of `class Peekable`, but, with another PR, we should be able to
complete issue #84 and make the lexer fully pull-driven by turning it
into a proper `Iterator`, which we would wrap within an instance of
`class Peekable`.

---------

Co-authored-by: Max Bernstein <max@bernsteinbear.com>